+++ redwood-flex-container/src/commonTest/kotlin/app/cash/redwood/flexcontainer/FlexContainerStringCanvasTest.kt
@@ -205,7 +205,7 @@ class FlexContainerStringCanvasTest {
       flexDirection = Row
       justifyContent = JustifyContent.Center
       items += widgets.map { widget ->
-        FlexItem(flexBasisPercent = 0f, measurable = widget).also { widget.flexItem = it }
+        FlexItem(flexBasisPercent = 0.0, measurable = widget).also { widget.flexItem = it }
       }
     }
 
@@ -276,11 +276,11 @@ class FlexContainerStringCanvasTest {
     widgets: List<StringWidget>,
   ): String {
     val canvas = StringCanvas(width, height)
-    val widthSpec = MeasureSpec.from(width, MeasureSpecMode.Exactly)
-    val heightSpec = MeasureSpec.from(height, MeasureSpecMode.Exactly)
+    val widthSpec = MeasureSpec.from(width.toDouble(), MeasureSpecMode.Exactly)
+    val heightSpec = MeasureSpec.from(height.toDouble(), MeasureSpecMode.Exactly)
 
     val result = measure(widthSpec, heightSpec)
-    layout(result, Size(width, height))
+    layout(result, Size(width.toDouble(), height.toDouble()))
 
     for (widget in widgets) {
       widget.draw(canvas)
